+++ src/hapi/pipelines/database/rainfall.py
@@ -4,7 +4,6 @@ from logging import getLogger
 from typing import Dict
 
 from hapi_schema.db_rainfall import DBRainfall
-from hdx.scraper.framework.utilities.reader import Read
 
 from .hapi_subcategory_uploader import HapiSubcategoryUploader
 
@@ -25,23 +24,8 @@ class Rainfall(HapiSubcategoryUploader):
         output_row["version"] = row["version"]
 
     def populate(self) -> None:
-        reader = Read.get_reader("hdx")
-        datasetinfo = {
-            "dataset": "hdx-hapi-rainfall",
-            "resource": "Global Climate: Rainfall datasets",
-            "format": "csv",
-        }
-        iso_list = []
-        headers, iterator = reader.read(datasetinfo=datasetinfo)
-        for row in iterator:
-            iso = row["location_code"].lower()
-            if "#" in iso:
-                continue
-            iso_list.append(iso)
-        for iso in iso_list:
-            # TODO: change end_resource to None when we're ready to load all years of data
-            self.hapi_populate(
-                f"rainfall-{iso}",
-                DBRainfall,
-                end_resource=1,
-            )
+        self.hapi_populate(
+            "rainfall",
+            DBRainfall,
+            end_resource=None,
+        )